requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic

########

pdfplumber
pillow
# pytesseract
tenacity